# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000/v1"
    description: "Server locale (v1)"
info:
  title: WASAText API
  description: Struttura API del progetto WASAText del corso di Programmazione Web.
  version: 1.0.0
...
security:
   - LegacyAuth: []
paths:
...
  /users/{userId}/chats:
    parameters:
        - $ref: "#/components/parameters/UserId"
    get:
      tags: ["chats"]
      summary: lista delle conversazioni di un utente
      description: lista delle conversazioni di un utente in ordine cronologico inverso o per ricerca per username
      operationId: getMyConversations
      parameters:
        - name: username
          in: query
          required: false
          description: parametri richiedenti lo userid e l'esistenza di esso
          schema:
            type: string
            pattern: '^.*?$'
            minLength: 3
            maxLength: 16
            description: ricerca una conversazione utilizzando un username (opzionale)
      responses:
        '200':
          description: lista conversazioni in ordine cronologico inverso ottenuta con successo (o risultato della ricerca)
          content:
            application/json:
              schema:
                type: object
                description: ritorna un oggetto con la lista "conversations" ordinata per ultimo messaggio.
                properties:
                  chats:
                    description: Elenco delle conversazioni trovate o esistenti
                    type: array
                    minItems: 0
                    maxItems: 100
                    items:
                      $ref: '#/components/schemas/Chat'
              example:
                    conversations:
                        - id: 10
                          isGroup: false
                          photoChat: "https://example.com/chat.jpg"
                          lastMessage: "2024-01-20T10:00:00Z"
                          snippetText: "Ciao!"
                          snippetIcon: "https://example.com/icon.png"

    post:
      tags: ["chats"]
      summary: crea una nuova conversazione o ritorna quella esistente
      description: crea una nuova conversazione uno-a-uno con un altro utente specificato tramite username, oppure ritorna quella esistente
      operationId: createConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Oggetto per la creazione di una nuova chat
              required:
                - username  
              properties:
                username:   
                  type: string
                  description: Username dell'altro utente con cui iniziare la conversazione
                  example: "Luigi"
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
      responses:
        "200":
          description: conversazione esistente ritornata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
              examples:
                existing:
                  value:
                    id: 10
                    isGroup: false
                    photoChat: "https://example.com/chat.jpg"
                    lastMessage: "2024-01-20T10:00:00Z"
                    snippetText: "Bentornato!"
                    snippetIcon: "https://example.com/icon.png"
        "201":
          description: nuova conversazione creata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
              examples:
                created:
                  value:
                    id: 11
                    isGroup: false
                    photoChat: "https://example.com/newchat.jpg"
                    lastMessage: "2024-01-21T10:00:00Z"
                    snippetText: "Nuova chat"
                    snippetIcon: "https://example.com/default.png"
        "400":
          description: richiesta non valida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: utente non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/chats/{chatId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/ChatId'
    get:
      tags: ["chats"]
      summary: messaggi di una conversazione
      description: messaggi di una conversazione in ordine cronologico inverso
      operationId: getConversation
      responses:
        '200':
          description: messaggi di una chat in ordine cronologico inverso ottenuti
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
              examples:
                success:
                  value:
                    id: 10
                    isGroup: false
                    photoChat: "https://example.com/chat.jpg"
                    lastMessage: "2024-01-20T10:00:00Z"
                    snippetText: "Ciao!"
                    snippetIcon: "https://example.com/icon.png"
                    messages:
                      - id: 46156161
                        text: "Ciao!"
                        sentBy: 123
                        sentAt: "2024-01-20T10:00:00Z"
                        checkmark: true
...
components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificatore numerico dell'utente
    ChatId:
      name: chatId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificatore numerico della chat
...
  schemas:
...
    Chat:
      type: object
      description: Dati sintetici di una conversazione per la lista
      properties:
        id:
          type: integer
          description: Identificatore della chat
        name:
          type: string
          description: Nome della chat 
          pattern: "^.*$"
        isGroup:
          type: boolean
          description: true se la conversazione è di gruppo
        photoChat:
          type: string
          format: url
          minLength: 10
          maxLength: 200
          description: URL dell'immagine rappresentativa della chat
        lastMessage:
          type: string
          format: date-time
          description: Timestamp dell'ultimo messaggio
        snippetText:
          type: string
          minLength: 1
          maxLength: 2000
          description: Anteprima del testo dell'ultimo messaggio
        snippetIcon:
          type: string
          format: url
          minLength: 10
          maxLength: 200
          description: Icona o immagine associata all'anteprima
        messages:
          description: Lista dei messaggi appartenenti alla conversazione
          type: array
          minItems: 0
          maxItems: 2000
          items:
            $ref: '#/components/schemas/Message'
    Message:
      type: object
      description: Rappresenta un messaggio inviato in una chat
      properties:
        id:
          type: integer
          description: Identificatore del messaggio
        text:
          type: string
          description: Testo del messaggio
          minLength: 0
          maxLength: 2000
        sentAt:
          type: string
          format: date-time
          description: Data/ora di invio
        sentBy:
          type: integer
          description: id utente che ha inviato
        checkmark:
          type: boolean
          description: stato di consegna/lettura
        senderName:
          type: string
          description: Nome visuale del mittente (per i gruppi)
        replyTo:
          type: integer
          description: ID del messaggio a cui questo risponde
        isForward:
          type: boolean
          description: True se il messaggio è inoltrato
        photoUrl:
          type: string
          description: URL della foto allegata
...
    Error:
      type: object
      description: schema di errore generico
      properties:
        code:
          type: integer
          description: Codice di errore
        message:
          type: string
          minLength: 5
          maxLength: 200
          pattern: "^.*$"
          description: messaggio di spiegazione dell'errore
  securitySchemes: 

    LegacyAuth:
      type: apiKey
      name: sessionToken
      in: header
      description: valore del token inviato dal server nel momento di login